# generated/example.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "example.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/graphql": {
      "get": {
        "operationId": "GetMarkersGraphQL",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/generatedGraphQlMarkersWrapper"
            }
          }
        },
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MarkersService"
        ]
      },
      "post": {
        "operationId": "GetMarkersGraphQL",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/generatedGraphQlMarkersWrapper"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/generatedGraphQLQuery"
            }
          }
        ],
        "tags": [
          "MarkersService"
        ]
      }
    },
    "/markers": {
      "get": {
        "operationId": "GetMarkers",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/generatedMarkers"
            }
          }
        },
        "tags": [
          "MarkersService"
        ]
      }
    }
  },
  "definitions": {
    "generatedGraphQLQuery": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string"
        }
      }
    },
    "generatedGraphQlMarkersWrapper": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/generatedMarkers"
        }
      }
    },
    "generatedMarker": {
      "type": "object",
      "properties": {
        "point": {
          "type": "integer",
          "format": "int32"
        },
        "homeTeam": {
          "type": "string"
        },
        "awayTeam": {
          "type": "string"
        },
        "markerImage": {
          "type": "string"
        },
        "information": {
          "type": "string"
        },
        "fixture": {
          "type": "string"
        },
        "capacity": {
          "type": "string"
        },
        "previousScore": {
          "type": "string"
        },
        "tv": {
          "type": "string"
        }
      }
    },
    "generatedMarkers": {
      "type": "object",
      "properties": {
        "markers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/generatedMarker"
          }
        }
      }
    }
  }
}